Refactor mullvad-nsis
* Separate NSIS exports from other code
* Share get_known_folder_path between mullvad-nsis and mullvad-paths
* Use PRODUCT_NAME in mullvad-nsis instead of literals
